@@ -30,10 +30,16 @@
     "apollo-cache-inmemory": "^1.6.5",
     "apollo-client": "^2.6.8",
     "apollo-link-http": "^1.5.16",
+    "d3": "^5.16.0",
+    "d3-render": "^0.2.3",
     "dotenv": "^8.2.0",
     "graphql": "^14.6.0",
     "graphql-tag": "^2.10.3",
+    "helmet": "^3.21.2",
+    "html-entities": "^1.3.1",
+    "http-proxy-middleware": "^0.20.0",
     "isomorphic-unfetch": "^3.0.0",
+    "leaflet": "^1.6.0",
     "lost": "^8.3.1",
     "next": "^9.3.1",
     "next-routes": "^1.4.2",
@@ -41,11 +47,13 @@
     "normalize.css": "^8.0.1",
     "prop-types": "^15.7.2",
     "query-string": "^6.10.1",
-    "react": "^16.12.0",
+    "react": "^16.13.1",
     "react-apollo": "^3.1.3",
-    "react-dom": "^16.12.0",
+    "react-dom": "^16.13.1",
     "react-gtm-module": "^2.0.8",
     "react-infinite-scroller": "^1.2.4",
+    "react-leaflet": "^2.7.0",
+    "react-leaflet-fullscreen": "^1.0.1",
     "react-photoswipe": "^1.3.0",
     "react-range": "^1.5.3",
     "react-redux": "^7.1.3",
@@ -54,6 +62,7 @@
     "reactronica": "^0.4.3",
     "redux": "^4.0.5",
     "stylelint-webpack-plugin": "^1.2.3",
+    "text-miner": "^1.3.0",
     "ua-compatible": "0.0.0",
     "vivus": "^0.4.5",
     "wpapi": "^1.2.1"
