@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import Link from '../Link';
 import TileImage from '../TileImage';
-import CTAButton from '../CTAButton';
+import CTALink from '../CTALink';
 import TileButtonGroup from '../TileButtonGroup';
 
 // import './MainTile.css';
@@ -97,7 +97,7 @@ class MainTile extends Component {
             <div className="main-tile__subtitle">{subtitle}</div>
             {showTitleButton && (
               <div className="main-tile__title-button">
-                <CTAButton>Launch Experiment</CTAButton>
+                <CTALink href={url}>Launch Experiment</CTALink>
               </div>
             )}
           </header>
